Added read-only roles and fixed file copy
