+++ src/deployments/cdk/src/deployments/defaults/step-1.ts
@@ -39,7 +39,6 @@ import { Account } from '../../utils/accounts';
 import { createDefaultS3Bucket, createDefaultS3Key } from './shared';
 import { overrideLogicalId } from '../../utils/cdk';
 import { getVpcSharedAccountKeys } from '../../common/vpc-subnet-sharing';
-import path from 'path';
 
 export type AccountRegionEbsEncryptionKeys = { [accountKey: string]: { [region: string]: kms.Key } | undefined };
 
@@ -68,9 +67,7 @@ export async function step1(props: DefaultsStep1Props): Promise<DefaultsStep1Res
   const logAccountDefaultKeys = createDefaultEncryptionKeys(props);
   const accountEbsEncryptionKeys = createDefaultEbsEncryptionKey(props);
   const aesLogBucket = createAesLogBucket(props);
-  if (props.config['global-options']['meta-data-collection']) {
-    createMetadataService(props);
-  }
+
   return {
     centralBucketCopy,
     centralLogBucket,
@@ -103,140 +100,6 @@ function blockS3PublicAccess(props: DefaultsStep1Props) {
 /**
  * Creates a bucket that contains copies of the files in the central bucket.
  */
-function createMetadataService(props: DefaultsStep1Props) {
-  const { accountStacks, config, acceleratorPrefix } = props;
-
-  const masterAccountConfig = config['global-options']['aws-org-management'];
-  const masterAccountStack = accountStacks.getOrCreateAccountStack(masterAccountConfig.account);
-
-  const organizations = new Organizations(masterAccountStack, 'MetadataOrganizations');
-
-  const keyAlias = createEncryptionKeyName('Metadata-Key');
-  const encryptionKey = new kms.Key(masterAccountStack, 'MetadataBucketKey', {
-    alias: `alias/${keyAlias}`,
-    description: 'Key used to encrypt/decrypt the metadata S3 bucket',
-    enableKeyRotation: true,
-  });
-
-  const bucket = new s3.Bucket(masterAccountStack, 'MetadataBucket', {
-    encryptionKey,
-    blockPublicAccess: s3.BlockPublicAccess.BLOCK_ALL,
-    removalPolicy: cdk.RemovalPolicy.RETAIN,
-    objectOwnership: s3.ObjectOwnership.BUCKET_OWNER_PREFERRED,
-    versioned: true,
-  });
-
-  // Let the bucket name be generated by CloudFormation
-  // The generated bucket name is based on the stack name + logical ID + random suffix
-  overrideLogicalId(bucket, `metadata${masterAccountStack.region}`);
-
-  const anyAccountPrincipal = [new iam.AnyPrincipal()];
-
-  // Give all accounts access to use this key for decryption
-  encryptionKey.addToResourcePolicy(
-    new iam.PolicyStatement({
-      actions: ['kms:Decrypt'],
-      principals: anyAccountPrincipal,
-      resources: ['*'],
-      conditions: {
-        StringEquals: {
-          'aws:PrincipalOrgID': organizations.organizationId,
-        },
-        ArnLike: {
-          'aws:PrincipalARN': `arn:aws:iam::*:role/${acceleratorPrefix}*`,
-        },
-      },
-    }),
-  );
-
-  // Give all accounts access to get and list objects in this bucket
-  bucket.addToResourcePolicy(
-    new iam.PolicyStatement({
-      actions: ['s3:Get*', 's3:List*'],
-      resources: [bucket.bucketArn, bucket.arnForObjects('*')],
-      principals: anyAccountPrincipal,
-      conditions: {
-        StringEquals: {
-          'aws:PrincipalOrgID': organizations.organizationId,
-        },
-        ArnLike: {
-          'aws:PrincipalARN': `arn:aws:iam::*:role/${acceleratorPrefix}*`,
-        },
-      },
-    }),
-  );
-
-  // Allow only https requests
-  bucket.addToResourcePolicy(
-    new iam.PolicyStatement({
-      actions: ['s3:*'],
-      resources: [bucket.bucketArn, bucket.arnForObjects('*')],
-      principals: anyAccountPrincipal,
-      conditions: {
-        Bool: {
-          'aws:SecureTransport': 'false',
-        },
-      },
-      effect: iam.Effect.DENY,
-    }),
-  );
-
-  const lambdaPath = require.resolve('@aws-accelerator/deployments-runtime');
-  const lambdaDir = path.dirname(lambdaPath);
-  const lambdaCode = lambda.Code.fromAsset(lambdaDir);
-  const lambdaRole = new iam.Role(masterAccountStack, 'metadata-lambda', {
-    assumedBy: new iam.ServicePrincipal('lambda.amazonaws.com'),
-    roleName: `${props.acceleratorPrefix}metadata-lambda`,
-  });
-
-  lambdaRole.addToPolicy(
-    new iam.PolicyStatement({
-      resources: ['*'],
-      actions: [
-        'codecommit:GetFile',
-        'dynamodb:BatchGetItem',
-        'dynamodb:GetRecords',
-        'dynamodb:GetShardIterator',
-        'dynamodb:Query',
-        'dynamodb:GetItem',
-        'dynamodb:Scan',
-        'kms:DescribeKey',
-        'kms:Decrypt',
-        'kms:Encrypt',
-        'kms:ReEncrypt*',
-        'kms:GenerateDataKey*',
-        'organizations:Describe*',
-        'organizations:List*',
-        'organizations:Get*',
-        'secretsmanager:Get*',
-        'ssm:Get*',
-        'states:List*',
-        'sts:GetCallerIndentity',
-      ],
-    }),
-  );
-
-  const metadataLambda = new lambda.Function(masterAccountStack, `MetadataLambda`, {
-    runtime: lambda.Runtime.NODEJS_14_X,
-    code: lambdaCode,
-    role: lambdaRole,
-    handler: 'index.metadataCollection',
-    timeout: cdk.Duration.minutes(5),
-    environment: {
-      ACCELERATOR_PREFIX: props.acceleratorPrefix,
-      BUCKET_NAME: bucket.bucketName,
-      CONFIG_REPOSITORY_NAME: process.env.CONFIG_REPOSITORY_NAME || `${props.acceleratorPrefix}Config-Repo`,
-    },
-  });
-
-  encryptionKey.grantEncrypt(metadataLambda);
-  bucket.grantWrite(metadataLambda);
-  const cloudwatchrule = new Rule(masterAccountStack, `MetadataCWRule`, {
-    schedule: Schedule.rate(cdk.Duration.days(1)),
-  });
-  cloudwatchrule.addTarget(new targets.LambdaFunction(metadataLambda));
-  return bucket;
-}
 
 /**
  * Creates a bucket that contains copies of the files in the central bucket.
